requirements.txt
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
Authlib==0.15.2
backcall==0.2.0
backoff==1.10.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
Django==3.1.2
django-crispy-forms==1.9.2
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-googlemaps==0.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Simple-GeoIP==0.2.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
ipaddress==1.0.23
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
lxml==4.6.1
MarkupSafe==1.1.1
nltk==3.5
packaging==20.4
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.7
py==1.9.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pyzmq==19.0.2
regex==2020.10.23
requests==2.24.0
selenium==3.141.0
simple-geoip==0.1.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.4
urllib3==1.25.11
uwsgi
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
